workouts: move WorkoutRowView to separate file
